reintroduce time words
