+++ Makefile.unix
@@ -29,6 +29,7 @@ INCLUDED_SRC=src/utils.c \
 	ext/ext-words.c \
     ext/parser-words.c \
 	ext/io-words.c \
+    ext/time-words.c \
 	ext/math-words.c
